serverless_functions/requirements.txt
azure-functions
requests
azure-storage-blob
xmltodict